[SCM-502] when checking out a tag we must do a git-fetch instead of a git-pull
This has do be done since the tag may be on another branch as the current
branch. Performing a git-pull failed in some cases witha 'cannot merge'
exception and if not it may have created an invalid merge in your local repository.

Also a tag in git may be on multiple branches. This happens if the branching
happens after the tag got created.

Checking out a tag will now create a 'detached HEAD' as specified by the
git documentation.